requirements.txt
torch
torchaudio
torchvision
faster-whisper
google-generativeai
python-dotenv
pydub
mutagen
demucs
soundfile
ollama
